# docs/swagger.yaml
basePath: /
definitions:
...
  auth.OAuthErrorResponse:
    properties:
      error:
        type: string
      error_description:
        type: string
    type: object
...
  clients.ClientCreateRequest:
    properties:
      allowed_scopes:
        type: string
      app_name:
        type: string
      domain:
        type: string
      logo_url:
        type: string
      owner_id:
        type: string
      redirect_uris:
        type: string
    type: object
  clients.ClientUpdateRequest:
    properties:
      allowed_scopes:
        type: string
      app_name:
        type: string
      domain:
        type: string
      is_active:
        type: boolean
      logo_url:
        type: string
      owner_id:
        type: string
      redirect_uris:
        type: string
    type: object
...
info:
  contact: {}
  description: SSO server API
  title: SSO API
  version: "1.0"
paths:
...
  /auth/{platform}/callback:
    get:
      parameters:
      - description: OAuth provider (discord or google)
        in: path
        name: platform
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: CSRF state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
      summary: OAuth callback
      tags:
      - auth
  /auth/{platform}/login:
    get:
      parameters:
      - description: OAuth provider (discord or google)
        in: path
        name: platform
        required: true
        type: string
      - description: Redirect URL after login
        in: query
        name: redirect_url
        type: string
      - description: OAuth flow ID (for server-side flow)
        in: query
        name: flow_id
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
      summary: OAuth login
      tags:
      - auth
...
  /auth/logout:
    get:
      parameters:
      - description: ID token hint
        in: query
        name: id_token_hint
        type: string
      - description: Post logout redirect URI
        in: query
        name: post_logout_redirect_uri
        type: string
      - description: State
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect
          schema:
            type: string
      summary: Logout
      tags:
      - auth
...
  /clients:
    get:
      description: List all OAuth clients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List OAuth clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Create a new OAuth client
      parameters:
      - description: Client creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clients.ClientCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create OAuth client
      tags:
      - clients
  /clients/{id}:
    get:
      description: Get details of a specific OAuth client
      parameters:
      - description: Client ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get OAuth client by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Update an existing OAuth client
      parameters:
      - description: Client ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Client update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clients.ClientUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update OAuth client
      tags:
      - clients
...
swagger: "2.0"